+++ Manuscript/Storage/BusinessModel/TaskBusinessModel.swift
@@ -11,18 +11,18 @@ import CoreData
 public struct TaskBusinessModel: BusinessModelProtocol {
     public var coreDataId: NSManagedObjectID? = nil
 
-    public let remoteId: Int32
+    public let remoteId: Int64
     public let title: String
     public var isInitiallySynced: Bool
     public let lastModifiedDate: Date
     public let detail: String?
     public let dueDate: String
-    public let ownerBoardId: Int32
+    public let ownerBoardId: Int64
     public let status: String
-    public let workspaceId: Int32
+    public let workspaceId: Int64
     public let isPendingDeletionOnTheServer: Bool
 
-    public init(remoteId: Int32, title: String, detail: String?, dueDate: String, ownerBoardId: Int32, status: String, workspaceId: Int32, lastModifiedDate: String, isInitiallySynced: Bool, isPendingDeletionOnTheServer: Bool) {
+    public init(remoteId: Int64, title: String, detail: String?, dueDate: String, ownerBoardId: Int64, status: String, workspaceId: Int64, lastModifiedDate: String, isInitiallySynced: Bool, isPendingDeletionOnTheServer: Bool) {
         self.remoteId = remoteId
         self.title = title
         self.detail = detail
