+++ Manuscript/Storage/Creators/BoardCreator.swift
@@ -63,7 +63,7 @@ class BoardCreator {
                 
                 context.performAndWait {
                     if let boardToBeUpdated = try? context.existingObject(with: coreDataId) as? BoardEntity {
-                        boardToBeUpdated.remoteId = Int32(boardResponse.id)
+                        boardToBeUpdated.remoteId = Int64(boardResponse.id)
                         boardToBeUpdated.lastModifiedDate = boardResponse.lastModifiedDate
                         boardToBeUpdated.isInitiallySynced = true
                         do {
